Bug 1795423 - Move Focus requestInPlacePermissions extension to AC support utils.
