+++ focus-android/app/src/main/java/org/mozilla/focus/fragment/BaseFragment.kt
@@ -8,7 +8,6 @@ import android.content.res.Resources
 import android.view.animation.Animation
 import android.view.animation.AnimationSet
 import android.view.animation.AnimationUtils
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.core.view.isInvisible
 import androidx.fragment.app.Fragment
 import org.mozilla.focus.ext.hideToolbar
@@ -54,19 +53,5 @@ abstract class BaseFragment : Fragment() {
         } else {
             null
         }
-    }
-}
-
-fun Fragment.requestInPlacePermissions(
-    permissionsToRequest: Array<String>,
-    onResult: (Map<String, Boolean>) -> Unit,
-) {
-    requireActivity().activityResultRegistry.register(
-        "permissionsRequest",
-        ActivityResultContracts.RequestMultiplePermissions(),
-    ) { permissions ->
-        onResult(permissions)
-    }.also {
-        it.launch(permissionsToRequest)
     }
 }
